Update LICENSE; Fix wrong returt type for divide_by_three function

+++ src/divide/divide_by_three.py
@@ -1,2 +1,2 @@
-def divide_by_three(num: int) -> int:
+def divide_by_three(num: int) -> float:
 	return num / 3
